+++ CaliberTrainerReport/src/app/Components/avg-qc-score-review/review-page-avg-qc-score.component.ts
