+++ CaliberTrainerReport/src/app/Components/qc-batch-scores/qc-overall-batch-technical-scores.component.ts
