+++ CaliberTrainerReport/src/app/Components/qc-dropdown-scores/qc-technical-scores-by-category-across-batches.component.ts
